fix: PointSaveData change x and y from float to int

+++ app/DataTransferObjects/Points/PointSaveData.php
@@ -7,8 +7,8 @@ use Spatie\DataTransferObject\DataTransferObject;
 class PointSaveData extends DataTransferObject
 {
     public string $name;
-    public float $x;
-    public float $y;
+    public int $x;
+    public int $y;
     public ?string $opened_at;
     public ?string $closed_at;
 }
